requirements.txt
aiogram>=3.0
aiohttp>=3.8.0
feedparser>=6.0.0
python-telegram-bot>=20.0
beautifulsoup4>=4.11.0
python-dotenv>=1.0.0